Prototype 3/agents_healthcheck/requirements.txt
blinker==1.8.2
click==8.1.7
cloudevents==1.10.1
confluent-kafka==2.4.0
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.6.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
watchdog==4.0.0
Werkzeug==3.0.3
